refactor: moved NetworkError to error.rs
