+++ crates/net/network-api/src/lib.rs
@@ -13,28 +13,10 @@ use async_trait::async_trait;
 use reth_primitives::{NodeRecord, H256, U256};
 use serde::{Deserialize, Serialize};
 use std::net::SocketAddr;
-use thiserror::Error;
-use tokio::sync::{mpsc, oneshot};
 
-/// Network Errors
-#[allow(missing_docs)]
-#[derive(Error, Debug, Clone, PartialEq, Eq)]
-pub enum NetworkError {
-    #[error("Sender has been dropped")]
-    ChannelClosed,
-}
-
-impl<T> From<mpsc::error::SendError<T>> for NetworkError {
-    fn from(_: mpsc::error::SendError<T>) -> Self {
-        NetworkError::ChannelClosed
-    }
-}
-
-impl From<oneshot::error::RecvError> for NetworkError {
-    fn from(_: oneshot::error::RecvError) -> Self {
-        NetworkError::ChannelClosed
-    }
-}
+/// Network Error
+pub mod error;
+pub use error::NetworkError;
 
 /// Provides general purpose information about the network.
 #[async_trait]
